refactor: Create Pipfile and remove requirements.txt
For a project that is actively being worked on, using pipenv instead of a
requirements.txt makes it much more convenient to manage dependencies, both
for development and for deployment. For this commit, the dependencies in
requirements.txt have been moved to a Pipfile as they were, even though not
all of the dependencies have even been made use of in the code yet.