test: Fix test_ten_percent_discount to actually test discount
Before this commit, test_ten_percent_discount would create a Discount
for toothbrushes, but then only add apples to the cart. With this
commit, toothbrushes are now added to the cart, so the test actually
checks if the discount is applied.
The way the discount prices are calculated and then checked in this test
is problematic for a number of reasons, but this will be changed in future
commits - this commit was only concerned with making sure that
test_ten_percent_discount actually tests what it promises.