refactor: Clean up functions in Teller
As part of this refactor, the ProductQuantity class is being removed from
model_objects.py. This is because the ProductQuantity instances, which have
previously only been created in the ShoppingCart class, are not actually needed
anywhere after all. This of course assumes that, in this refactoring task for
this theoretical app, the ProductQuantity entities are not expected to be
written into the database for BI purposes.