refactor: Refactor code for discount creation
The code for Discount creation in the ShoppingCart class was a travesty; it
was extremely hard to read and to debug. It also arguably shouldn't have been
in the ShoppingCart in the first place, but rather in a separate module.
With this commit, the Discount creation code has been moved into a new
discount_creation module, and the logic that calculates the Discounts is much
more clean.