feat: Add validation for optional argument of Offer
The argument (now renamed to optional_argument) in the Offer class is only
needed if the Offer has one of some specific offer types. The optional_argument
now allows for None values, and new validation makes sure that, when a Discount
is created from an Offer and the Offer has None for the optional_argument
instead of the expected float, a ValueError is raised.