test: Add test to cover that prices in receipt can be 0
It's currently possible to have receipts with items on them where the total
price comes out at 0, either because the bought quantities are lower than one
cent's worth (e.g. one gram of apples) or because a high discount has been
applied. This could easily be changed in the get_total_price_cents method in
Receipt, e.g. be returning max(total, 1). However, since this changes the
business logic, such a change is not being made as part of this refactoring
project.