feat: Make sure Product is in Catalog before adding to ShoppingCart
Before this commit, ShoppingCarts were not associated with any specific
Catalog, even though the Teller that would later check out the ShoppingCart is.
Now, a ShoppingCart belongs to a specific Catalog, and a Product can only be
added to the ShoppingCart if it is included in that Catalog as well.